Easy mode lag bug fixed
In easy mode the snake sometimes did not move when the directional button was pressed. We suspect this issue is because of a difference in the refresh rate and speed of movement in the game. By loweing the GameTimer.Interval for all three game modes the bug was eliminated and the snake does not lag on any difficulty setting.